[VISUALIZE] Create sidebar and reorganize editor. closes #16
