+++ app/components/editor/editor-canvas/editor-canvas.module.scss
@@ -1,0 +1,10 @@
+.editor-layout-canvas {
+  position: relative;
+  max-height: 100%;
+  height: 100%;
+  cursor: grab;
+
+  &.dragging {
+    cursor: grabbing;
+  }
+}
