+++ app/components/editor/editor-cursor/editor-cursor.module.scss
@@ -7,6 +7,7 @@
   border-left: 1px solid blue;
   padding: 5px;
   min-width: 70px;
+  z-index: 1;
 
   &-position {
     &-name {
