+++ app/components/editor/editor-sidebar/editor-sidebar.module.scss
@@ -1,0 +1,10 @@
+.sidebar-container {
+  width: 10rem;
+  background-color: white;
+  border-right: 1px solid black;
+  position: absolute;
+  left: 0;
+  top: 0;
+  height: 100%;
+  z-index: 1;
+}
