+++ app/components/editor/editro.module.scss
@@ -3,14 +3,14 @@
   display: flex;
   flex-direction: column;
   position: relative;
-  .paper-canvas {
-    width: 100% !important;
-    height: 100% !important;
-    flex: 1;
-    cursor: grab;
+  width: 100%;
+  height: 100%;
 
-    &.dragging {
-      cursor: grabbing;
-    }
+  .editor-layout {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+    height: 100%;
+    width: 100%;
   }
 }
