+++ app/convert/visualize/styles.module.scss
@@ -1,5 +1,5 @@
 .visualize {
-  min-height: 100vh;
+  height: 100vh;
   display: flex;
   flex-direction: column;
 
@@ -8,5 +8,7 @@
     display: flex;
     flex-direction: column;
     border: 1px solid black;
+    height: 100%;
+    min-height: 0px;
   }
 }
